Add anime prompt builder with tab navigation
